Excluir celular funcionando

+++ app/Http/Controllers/CelularController.php
@@ -196,8 +196,23 @@ class CelularController extends Controller
      */
     public function destroy($id)
     {
+        // Verifica se usuário tem permissões de acesso
+        $user = User::find(auth()->user()->id);
+        if (!$user->can('gerenciar celulares')){
+            return response()->json([
+                'success' => false,
+                'errors' => ['Você não possui permissão para realizar essa ação.'],
+            ])->setStatusCode(201);
+        }
+        
         $celular = Celular::find($id);
+        if (!isset($celular)) return abort(404);
+
         $celular->delete();
-        return view('celular.index');
+        
+        return response()->json([
+            'success' => true,
+            'message' => 'Celular excluído com sucesso',
+        ]);
     }
 }
